Deduplicate enum test classes and remove an obsolete test
